% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_CPTAC_ExpCounts.R
\docType{data}
\name{CPTAC_exp_counts}
\alias{CPTAC_exp_counts}
\title{Peptides in the CPTAC database}
\format{
A data frame with 1560773 rows and 2 variables:
\describe{
  \item{sequence}{amino acid sequence of peptide}
  \item{n_obs_pep}{number of experiments peptide was observed in}

}
}
\usage{
CPTAC_exp_counts
}
\description{
A data frame containing peptides found in a 2021/02/18 download of the CPTAC database and the number of experiments that they were observed in.
}
\keyword{datasets}
