% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_peptidome_generator.R
\name{create_peptidome}
\alias{create_peptidome}
\title{Peptidome Generator}
\usage{
create_peptidome(proteome_dir, missed_cleavages, synth_peps, aa_range)
}
\arguments{
\item{proteome_dir}{directory to proteome fasta file}

\item{missed_cleavages}{range of missed cleavage events considered during in-silico digestion}

\item{synth_peps}{If TRUE, filters out peptides that are problematic for peptide synthesis. Currnetly includes peptides with: N-term Q/E residues and C-term P residues}

\item{aa_range}{range of peptide size by amino acid count included}
}
\value{
Peptidome of unique peptides with parent protein uniprot IDs, sequence locations in parent proteins, and n missed cleavages
}
\description{
Generates a tryptic peptidome of unique peptides using a provided proteome. Requires bioconductor pacakge "cleaver".
}
