% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_peptide_predictor.R
\name{peptide_predictions}
\alias{peptide_predictions}
\title{Peptide Detectability Predictor}
\usage{
peptide_predictions(peptides, prediction_model, missed_cleavages)
}
\arguments{
\item{peptides}{list of peptide amino acid sequences of interest.}

\item{prediction_model}{RF model for used for detectability predictions.}

\item{missed_cleavages}{list of peptide missed cleavages in same order as peptides. If missing, missed cleavages will be calculated from peptide sequence for model predictions.}
}
\value{
dataframe of RF detectability scores with peptide sequences.
}
\description{
Predicts detectability scores for given peptide(s) using provided RF model.
}
\examples{
\dontrun{
peptide_predictions(peptides = c("PEPTIDERANGER", "LLYQPILR", "DAASLFGNK"),
                    prediction_model = RFmodel_ProteomicsDB)

peptide_predictions(peptides = "MLAILKAQILLR",
                    prediction_model = RFmodel_CPTAC,
                    missed_cleavages = 2)
}

}
