% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_peptides_inReference.R
\name{peptides_inReference}
\alias{peptides_inReference}
\title{Peptide Reference Merger}
\usage{
peptides_inReference(
  peptidome,
  pep_reference,
  reference_name,
  exp_counts_col,
  detection_freq,
  detection_ratio
)
}
\arguments{
\item{peptidome}{list of all possible peptides for proteins of interest and their uniprot IDs}

\item{pep_reference}{dataframe with a "sequence" column of peptides and a column of n observations in a reference}

\item{reference_name}{the name of the reference, which is used to label the output columns}

\item{exp_counts_col}{the name of the reference column containing the n of peptide observations}

\item{detection_freq}{If TRUE, calculates the each peptide's frequency of appearance in the reference}

\item{detection_ratio}{If TRUE, calculates ratio of peptide observations to parent protein observations}
}
\value{
dataframe of peptidome with reference information
}
\description{
Matches peptide reference information with peptidome sequences
}
\examples{
\dontrun{

peptides_inReference(peptidome = SwissProt2018_peptidome,
                     pep_reference = tm_peptides,
                     reference_name = "transmembrane")

peptides_inReference(peptidome = SwissProt2018_peptidome,
                     pep_reference = CPTAC_exp_counts,
                     reference_name = "CPTAC",
                     exp_counts_col = "n_obs_pep",
                     detection_freq = TRUE,
                     detection_ratio = TRUE)
}


}
