% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_peptide_prioritizer.R
\name{prioritize_peptides}
\alias{prioritize_peptides}
\title{Peptide Prioritizer}
\usage{
prioritize_peptides(
  uniprot_list,
  max_n,
  peptidome,
  prediction_model,
  priorities,
  priority_thresholds
)
}
\arguments{
\item{uniprot_list}{a list of uniprot IDs for proteins of interest}

\item{max_n}{max number of peptides selected for each protein}

\item{peptidome}{list of all possible peptides for relevant proteome, their uniprot IDs, and the values for their non-RF priorities}

\item{prediction_model}{RF model used to predict scores for peptide detectability}

\item{priorities}{list of priority column names in peptidome}

\item{priority_thresholds}{list of minimum values that priorities must reach to be considered}
}
\value{
list of n prioritized peptides for proteins of interest
}
\description{
Selects the top n peptides for a list of protein uniprot IDs given a set of priorities
}
