% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ProtDB_ExpCounts.R
\docType{data}
\name{ProtDB_exp_counts}
\alias{ProtDB_exp_counts}
\title{Peptides in the ProteomicsDB database}
\format{
A data frame with 939371 rows and 2 variables:
\describe{
  \item{sequence}{amino acid sequence of peptide}
  \item{n_obs_pep}{number of experiments peptide was observed in}

}
}
\usage{
ProtDB_exp_counts
}
\description{
A dataset containing peptides found in the ProteomicsDB database and the number of experiments that they were observed in.
}
\keyword{datasets}
