% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_CPTAC_RF.R
\docType{data}
\name{RFmodel_CPTAC}
\alias{RFmodel_CPTAC}
\title{CPTAC Trained Random Forest}
\format{
A random forest model trained using the randomForest package
}
\usage{
RFmodel_CPTAC
}
\description{
Random forest trained on the CPTAC database to predict peptide detectability.
}
\keyword{datasets}
