% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ProtDB_RF.R
\docType{data}
\name{RFmodel_ProteomicsDB}
\alias{RFmodel_ProteomicsDB}
\title{ProteomicsDB Trained Random Forest}
\format{
A random forest model trained using the randomForest package
}
\usage{
RFmodel_ProteomicsDB
}
\description{
Random forest trained on the ProteomicsDB database to predict peptide detectability.
}
\keyword{datasets}
