% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_tmPeptides.R
\docType{data}
\name{tm_peptides}
\alias{tm_peptides}
\title{List of Transmembrane Peptides}
\format{
A data frame with 14121 rows and 6 variables:
\describe{
  \item{symbol}{gene symbol of parent protein}
  \item{uniprot}{uniprot ID of parent protein}
  \item{sequence}{amino acid sequence of peptide}
  \item{size}{number of amino acid residues in peptide}
  \item{start}{parent protein position of first residue in peptide}
  \item{end}{parent protein position of last residue in peptide}

}
}
\usage{
tm_peptides
}
\description{
List of peptides that overlap with transmembrane domain sequence annotations downloaded from SwissProt-UniProt on 2021/01/14.
}
\keyword{datasets}
