% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_RFmodel_trainer.R
\name{train_RFmodel}
\alias{train_RFmodel}
\title{RF Model Training}
\usage{
train_RFmodel(peptidome, reference_name, total_trainingSet)
}
\arguments{
\item{peptidome}{List of all possible peptide sequences for relevant proteome (total_trainingSet = FALSE) OR a list of peptide sequences in a custom training set (total_trainingSet = TRUE). Must also include peptide n of observations (n_obs_pep) and parent protein n of observations (n_obs_prot) in a desired reference, and peptide missed cleavages.}

\item{reference_name}{Name of reference (same as column names)}

\item{total_trainingSet}{If TRUE, model will be trained on all peptides in a custom input peptidome. If FALSE, peptides will be selected from whole input peptidome to create a training set of 10k peptides (same method as models provided in package).}
}
\value{
RF model for predicting peptide detectability
}
\description{
Trains a random forest model using a peptidome with reference information
}
\examples{
\dontrun{
CPTAC_peptidome <- peptides_inReference(peptidome = SwissProt2018_peptidome,
                                        reference_name = "CPTAC",
                                        pep_reference = CPTAC_exp_counts,
                                        exp_counts_col = "n_obs_pep",
                                        detection_ratio = TRUE)

train_RFmodel(peptidome = CPTAC_peptidome,
              reference_name = "CPTAC")
}

}
