% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_peptide_prioritizer.R
\name{prioritize_peptides}
\alias{prioritize_peptides}
\title{Peptide Prioritizer}
\usage{
prioritize_peptides(
  uniprot_list,
  max_n,
  peptidome,
  prediction_model,
  priorities,
  priority_thresholds
)
}
\arguments{
\item{uniprot_list}{a list of uniprot IDs for proteins of interest}

\item{max_n}{max number of peptides selected for each protein}

\item{peptidome}{list of all possible peptides for relevant proteome, their uniprot IDs, and the values for their non-RF priorities}

\item{prediction_model}{RF model used to predict scores for peptide detectability}

\item{priorities}{list of priority column names in peptidome}

\item{priority_thresholds}{list of minimum values that priorities must reach to be considered}
}
\value{
list of n prioritized peptides for proteins of interest
}
\description{
Selects the top n peptides for a list of protein uniprot IDs given a set of priorities
}
\examples{
\dontrun{
prioritize_peptides(uniprot_list = c("Q9NQ94", "P04217", "A8K2U0"),
                    peptidome = SwissProt2018_peptidome,
                    prediction_model = CPTAC_RFmodel,
                    priorities = "RF_score",
                    priority_thresholds = 0)


CPTAC_peptidome <- peptides_inReference(peptidome = SwissProt2018_peptidome,
                                        reference_name = "CPTAC",
                                        pep_reference = CPTAC_exp_counts,
                                        exp_counts_col = "n_obs_pep",
                                        detection_freq = TRUE)

prioritize_peptides(uniprot_list = c("Q3T906", "P04217", "A8K2U0"),
                    max_n = 10, peptidome = CPTAC_peptidome,
                    prediction_model = RFmodel_CPTAC,
                    priorities = c("CPTAC_freq","RF_score"),
                    priority_thresholds = c(0.8, 0))
}


}
